+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.3] - Mar-31-2021
+
+**Milestone**: Mainnet(1.0.0.0)
+
+| Package          | Version | Link                                                               |
+| ---------------- | ------- | ------------------------------------------------------------------ |
+| Symbol Bootstrap | v1.0.3  | [symbol-bootstrap](https://www.npmjs.com/package/symbol-bootstrap) |
+
+-   Improved Custom Preset Object types for symbol bootstrap lib integration.
+-   TransactionSelectionStrategy's new default value is `oldest`.
+
 ## [1.0.2] - Mar-24-2021
 
 **Milestone**: Mainnet(1.0.0.0)